+++ pmc_automation_tools/driver/ux/driver.py
@@ -281,11 +281,7 @@ class UXDriver(PlexDriver):
             if column_match is None:
                 raise GridColumnError(f'No column detected in the table matching provided value: {column}.')
             column = column_match
-        rows = self.driver.find_elements(By.XPATH, f"//tr[contains(@class,'plex-grid-row selectable')]/td[@data-col-index={column}]")
-        matching_rows = []
-        for r in rows:
-            if r.get_property('textContent') == value:
-                matching_rows.append(r)
+        matching_rows = self.driver.find_elements(By.XPATH, f"//tr[contains(@class,'plex-grid-row selectable')]/td[@data-col-index={column} and text()='{value}']")
         if len(matching_rows) == 0:
             raise GridRowError(f"Plex grid row not found for column index {column} containing value: {value}.")
         if len(matching_rows) > 1: