+++ CHANGELOG.md
@@ -26,6 +26,12 @@ Added function to find an element using the element's text label. `find_element_
 
 Added support to `UXPlexElement` for universal `sync()` method when using above function to locate the element.
 
+Added function aliases for `wait_for_gears()` and `wait_for_banner()` to UXDriver class.
+
+## Changed
+
+Added support for `sync_checkbox()` input to accept multiple options for truthiness. True, 1, "1", "True", "TRUE" should all evaluate to True.
+
 # 0.6.3 [2025-1-28]
 
 ## Fixed